fix(ci): fix docker job (#182)
**Motivation**

The CI is broken on main. The job of building and publishing the docker
image is the cause of the failure.

**Description**

- Add `Dockerfile` that was missing for building the docker image.
- Remove the `[[example]]` part on the `Cargo.toml`. This is because
cargo already detect that `simple_usage.rs` is an example and we are
defining it with the same name, causing a duplicated error:
  ```
  Caused by:
found duplicate example name simple_usage, but all example targets must
have a unique name
  ```

A successful run was made
[here](https://github.com/lambdaclass/rex/actions/runs/16570331318/job/46860556506)