feat(ci,sdk): add rex daily report (#176)
**Motivation**

We need a way to know when rex is outdated from the ethrex repository.
<!-- Why does this pull request exist? What are its goals? -->

**Description**

A daily report was created, using the integration tests of the ci.
Instead of pointing to a specific commit from ethrex, we use the latest
version to test if rex is outdated.

After running the jobs, sends a slack message to a channel.

<!-- A clear and concise general description of the changes this PR
introduces -->

<!-- Link to issues: Resolves #111, Resolves #222 -->


Closes #issue_number

---------

Co-authored-by: ilitteri <ilitteri@fi.uba.ar>