refactor(sdk): remove duplicated code with ethrex (#192)
**Motivation**

We have a lot of code duplicated in both ethrex and rex, which adds
complexity and makes the code harder to maintain.

**Description**

- Reuses ethrex's `EthClient`.
- Removes the `rex::sdk` functions that are exact copies of those in
ethrex and reuses the ethrex versions throughout the codebase instead.
- The functions that remain in `rex::sdk` were kept because they differ
slightly from the ethrex implementations, for example, they allow
setting a custom `common_bridge_address` or other extra parameters.
These functions could also be deleted, but that would first require
modifying the ethrex functions to avoid losing functionality.

Related to #185